json-writer: fixups for V5
Please squash this onto the top of jh/json-writer.

Fix leading whitespace in t0019 using tricked suggested by Junio.
Fix unnecessary cast for intmax_t suggested by Wink.

+++ json-writer.c
@@ -165,7 +165,7 @@ void jw_object_string(struct json_writer *jw, const char *key, const char *value
 void jw_object_intmax(struct json_writer *jw, const char *key, intmax_t value)
 {
 	object_common(jw, key);
-	strbuf_addf(&jw->json, "%"PRIdMAX, (intmax_t)value);
+	strbuf_addf(&jw->json, "%"PRIdMAX, value);
 }
 
 void jw_object_double(struct json_writer *jw, const char *key, int precision,
@@ -303,7 +303,7 @@ void jw_array_string(struct json_writer *jw, const char *value)
 void jw_array_intmax(struct json_writer *jw, intmax_t value)
 {
 	array_common(jw);
-	strbuf_addf(&jw->json, "%"PRIdMAX, (intmax_t)value);
+	strbuf_addf(&jw->json, "%"PRIdMAX, value);
 }
 
 void jw_array_double(struct json_writer *jw, int precision, double value)
